+++ exporter/opentelemetry-exporter-jaeger/CHANGELOG.md
@@ -5,6 +5,9 @@
 - Change package name to opentelemetry-exporter-jaeger
   ([#953](https://github.com/open-telemetry/opentelemetry-python/pull/953))
 
+- Thrift URL for Jaeger exporter doesn't allow HTTPS (hardcoded to HTTP)
+  ([#978] (https://github.com/open-telemetry/opentelemetry-python/pull/978))
+
 ## 0.8b0
 
 Released 2020-05-27